Add protopath parser.
This adds a scanner/parser combination that is able to parse the same
output of a Path's string rendering. It also allows parsing a path with
an implicit root when the message type is known.

When you have a path in hand, it is also useful to retrieve the values
from a message along that path, so this change adds PathValues for
producing a protopath.Values for a given path and message.

This has been fuzzed internally without any findings.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>