Add Merge to prototext.UnmarshalOptions
This adds a Merge option to prototext.UnmarshalOptions to match
proto.UnmarshalOptions.  This is particularly useful when needing to
read multiple textproto files of the same format that wrap a single
repeated field.

I've added this to the exact same position in the UnmarshalOptions
struct as the proto version.  It can be moved to the bottom of the
struct definition if that would cause backwards compatibility issues,
but I wasn't aware of any.