proto: un-flake TestHasExtensionNoAlloc
Before this change, this test had a flakiness rate of 3 out of 1000 runs
(on Google’s internal testing infrastructure).

After this change, the test is not flaky, even when running with 50_000 runs.

testing.AllocsPerRun is meant to be robust against individual allocations
from background goroutines, but not in the way we called the function.

Change-Id: If99573734fb2de5d7128b5e61c2019867ccbb9b2
Reviewed-on: https://go-review.googlesource.com/c/protobuf/+/716520
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Christian Höppner <hoeppi@google.com>