encoding/prototext: Support URL chars in type URLs in text-format.
Change the text-format parser to allow for more characters and formats
in the type URL prefixes of expanded Any protos. This follows a recent
change to the text-format spec [1,2] and the Go implementation is now
almost exactly implementing this spec (with some exceptions to ensure
backwards compatibility).

Note that we are also making the whitespace handling between [ and ]
more lenient compared to the current implementation. This is also
in line with the new spec.

Refs:
- [1] https://protobuf.dev/reference/protobuf/textformat-spec/#characters
- [2] https://protobuf.dev/reference/protobuf/textformat-spec/#field-names
- go/protobuf-improved-type-url-support

Change-Id: I55055becec0248c1d161e776f1937eaaa4af2066
Reviewed-on: https://go-review.googlesource.com/c/protobuf/+/726780
Reviewed-by: Michael Stapelberg <stapelberg@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Damien Neil <dneil@google.com>